requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap4==2022.1
cryptography==38.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.2
django-allauth==0.52.0
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-embed-video==1.4.8
django-etc==1.4.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-social-share==2.3.0
django-taggit==3.1.0
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
pbr==5.11.1
Pillow==9.2.0
platformdirs==3.0.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stevedore==5.0.0
tzdata==2022.5
urllib3==1.26.13
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
